Image container add bmap to sparse images
Using dd when writing sparse images to non-zeroed disks sometimes resulted in
data corruption depending on previously written data along with the impression
of results being dependent on planetary alignment.
Understanding the intrinsics and implementation details of dd is difficult
and instead of diving deeper in that direction another tool is taken into use.

bmaptool is designed specifically for this use case and generates "bmaps"
or block maps of, typically, sparse full-disk images.
The block maps are used as an instruction for which part of a sparse image
to write during installation.

This patch will generate block maps when creating containers and use them
by calling image-install with the image type "raw-bmap".

Note: bmaptool will automatically find bmap files when named "IMAGE.bmap".